+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/impl/ClassRegistry.java
@@ -111,7 +111,7 @@ public abstract class ClassRegistry {
         public final StaticObject classData;
         public final boolean isHidden;
         public final boolean isStrongHidden;
-        public int klassID = -1;
+        public long klassID = -1;
 
         public boolean addedToRegistry() {
             return !isAnonymousClass() && !isHidden();
@@ -137,7 +137,7 @@ public abstract class ClassRegistry {
             return flags;
         }
 
-        public void initKlassID(int futureKlassID) {
+        public void initKlassID(long futureKlassID) {
             this.klassID = futureKlassID;
         }
     }
@@ -208,7 +208,7 @@ public abstract class ClassRegistry {
         }
     }
 
-    private final int loaderID;
+    private final long loaderID;
 
     private ModuleEntry unnamed;
     private final PackageTable packages;
@@ -218,7 +218,7 @@ public abstract class ClassRegistry {
         return unnamed;
     }
 
-    public final int getLoaderID() {
+    public final long getLoaderID() {
         return loaderID;
     }
 
@@ -246,7 +246,7 @@ public abstract class ClassRegistry {
      */
     private volatile Collection<Klass> strongHiddenKlasses = null;
 
-    protected ClassRegistry(int loaderID) {
+    protected ClassRegistry(long loaderID) {
         this.loaderID = loaderID;
         ReadWriteLock rwLock = new ReentrantReadWriteLock();
         this.packages = new PackageTable(rwLock);
