+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/impl/Klass.java
@@ -459,14 +459,14 @@ public abstract class Klass implements ModifiersProvider, ContextAccess, KlassRe
     static final Comparator<Klass> KLASS_ID_COMPARATOR = new Comparator<>() {
         @Override
         public int compare(Klass k1, Klass k2) {
-            return Integer.compare(k1.id, k2.id);
+            return Long.compare(k1.id, k2.id);
         }
     };
 
     static final Comparator<ObjectKlass.KlassVersion> KLASS_VERSION_ID_COMPARATOR = new Comparator<>() {
         @Override
         public int compare(ObjectKlass.KlassVersion k1, ObjectKlass.KlassVersion k2) {
-            return Integer.compare(k1.getKlass().getId(), k2.getKlass().getId());
+            return Long.compare(k1.getKlass().getId(), k2.getKlass().getId());
         }
     };
 
@@ -481,7 +481,7 @@ public abstract class Klass implements ModifiersProvider, ContextAccess, KlassRe
     protected Symbol<Type> type;
     private final EspressoContext context;
 
-    private final int id;
+    private final long id;
 
     @CompilationFinal //
     private volatile ArrayKlass arrayClass;
@@ -597,7 +597,7 @@ public abstract class Klass implements ModifiersProvider, ContextAccess, KlassRe
         this(context, name, type, modifiers, -1);
     }
 
-    Klass(EspressoContext context, Symbol<Name> name, Symbol<Type> type, int modifiers, int possibleID) {
+    Klass(EspressoContext context, Symbol<Name> name, Symbol<Type> type, int modifiers, long possibleID) {
         this.context = context;
         this.name = name;
         this.type = type;
@@ -844,7 +844,7 @@ public abstract class Klass implements ModifiersProvider, ContextAccess, KlassRe
         return fastLookup(this, interfaces) >= 0;
     }
 
-    public final int getId() {
+    public final long getId() {
         return id;
     }
