+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/impl/LoadingConstraints.java
@@ -60,8 +60,8 @@ import com.oracle.truffle.espresso.runtime.StaticObject;
 final class LoadingConstraints implements ContextAccess {
     private static DebugTimer CONSTRAINTS = DebugTimer.create("constraints");
 
-    private static final int NULL_KLASS_ID = -1;
-    static final int INVALID_LOADER_ID = -1;
+    private static final long NULL_KLASS_ID = -1;
+    static final long INVALID_LOADER_ID = -1;
 
     private final EspressoContext context;
     private final PurgeInfo info = new PurgeInfo();
@@ -87,8 +87,8 @@ final class LoadingConstraints implements ContextAccess {
      * Records that loader resolves type as klass.
      */
     void recordConstraint(Symbol<Type> type, Klass k, StaticObject loader) {
-        int loaderID = getLoaderID(loader, getMeta());
-        int klass = getKlassID(k);
+        long loaderID = getLoaderID(loader, getMeta());
+        long klass = getKlassID(k);
         ConstraintBucket bucket = lookup(type);
         if (bucket == null) {
             bucket = new ConstraintBucket();
@@ -115,8 +115,8 @@ final class LoadingConstraints implements ContextAccess {
     }
 
     void removeUnloadedKlassConstraint(Klass klass, Symbol<Type> type) {
-        int loaderId = getLoaderID(klass.getDefiningClassLoader(), getMeta());
-        int klassId = getKlassID(klass);
+        long loaderId = getLoaderID(klass.getDefiningClassLoader(), getMeta());
+        long klassId = getKlassID(klass);
         ConstraintBucket bucket = lookup(type);
         Constraint toRemove = bucket.lookupLoader(loaderId);
         bucket.remove(toRemove);
@@ -127,7 +127,7 @@ final class LoadingConstraints implements ContextAccess {
     }
 
     void purge() {
-        int[] alive = context.getRegistries().aliveLoaders();
+        long[] alive = context.getRegistries().aliveLoaders();
         info.emptyBuckets = 0;
         for (ConstraintBucket bucket : pairings.values()) {
             synchronized (bucket) {
@@ -148,11 +148,11 @@ final class LoadingConstraints implements ContextAccess {
                         "empty buckets: " + info.emptyBuckets);
     }
 
-    private void checkOrAdd(Symbol<Type> type, int k1, int k2, int loader1, int loader2) {
+    private void checkOrAdd(Symbol<Type> type, long k1, long k2, long loader1, long loader2) {
         if (exists(k1) && exists(k2) && k1 != k2) {
             throw linkageError("Loading constraint violated !");
         }
-        int klass = !exists(k1) ? k2 : k1;
+        long klass = !exists(k1) ? k2 : k1;
         ConstraintBucket bucket = lookup(type);
         if (bucket == null) {
             bucket = new ConstraintBucket();
@@ -196,7 +196,7 @@ final class LoadingConstraints implements ContextAccess {
 
         private Constraint constraint;
 
-        Constraint lookupLoader(int loader) {
+        Constraint lookupLoader(long loader) {
             Constraint curr = constraint;
             while (curr != null) {
                 if (curr.contains(loader)) {
@@ -207,7 +207,7 @@ final class LoadingConstraints implements ContextAccess {
             return null;
         }
 
-        Constraint lookupKlass(int klass) {
+        Constraint lookupKlass(long klass) {
             Constraint curr = constraint;
             while (curr != null) {
                 if (curr.klass == klass) {
@@ -241,7 +241,7 @@ final class LoadingConstraints implements ContextAccess {
             }
         }
 
-        void purge(int[] alive, PurgeInfo info) {
+        void purge(long[] alive, PurgeInfo info) {
             assert Thread.holdsLock(this);
             Constraint curr = constraint;
             while (curr != null) {
@@ -256,21 +256,21 @@ final class LoadingConstraints implements ContextAccess {
     }
 
     private static final class Constraint {
-        private int klass;
+        private long klass;
 
         /*
          * Most applications will only see the boot loader and the app class loader used.
          */
         private static final int DEFAULT_INITIAL_SIZE = 2;
 
-        private int[] loaders = new int[DEFAULT_INITIAL_SIZE];
+        private long[] loaders = new long[DEFAULT_INITIAL_SIZE];
         private int size = 0;
         private int capacity = DEFAULT_INITIAL_SIZE;
 
         Constraint prev;
         Constraint next;
 
-        boolean contains(int loader) {
+        boolean contains(long loader) {
             for (int i = 0; i < size; i++) {
                 if (loaders[i] == loader) {
                     return true;
@@ -280,18 +280,18 @@ final class LoadingConstraints implements ContextAccess {
         }
 
         void merge(Constraint other) {
-            for (int loader : other.loaders) {
+            for (long loader : other.loaders) {
                 if (!contains(loader)) {
                     add(loader);
                 }
             }
         }
 
-        Constraint(int k) {
+        Constraint(long k) {
             this.klass = k;
         }
 
-        void add(int loader) {
+        void add(long loader) {
             assert !contains(loader);
             if (size >= capacity) {
                 loaders = Arrays.copyOf(loaders, capacity <<= 1);
@@ -299,7 +299,7 @@ final class LoadingConstraints implements ContextAccess {
             loaders[size++] = loader;
         }
 
-        public void purge(int[] alive, PurgeInfo info) {
+        public void purge(long[] alive, PurgeInfo info) {
             int i = 0;
             while (i < size) {
                 if (!isAlive(loaders[i], alive)) {
@@ -317,21 +317,21 @@ final class LoadingConstraints implements ContextAccess {
                 while (size < (capacity >> (shift + 1))) {
                     shift += 1;
                 }
-                int[] newLoaders = new int[capacity = Math.max(DEFAULT_INITIAL_SIZE, capacity >> shift)];
+                long[] newLoaders = new long[capacity = Math.max(DEFAULT_INITIAL_SIZE, capacity >> shift)];
                 System.arraycopy(loaders, 0, newLoaders, 0, size);
                 loaders = newLoaders;
             }
         }
 
-        static void swap(int i, int j, int[] loaders) {
-            int a = loaders[i];
+        static void swap(int i, int j, long[] loaders) {
+            long a = loaders[i];
             loaders[i] = loaders[j];
             loaders[j] = a;
         }
 
-        static boolean isAlive(int loader, int[] alive) {
+        static boolean isAlive(long loader, long[] alive) {
             for (int i = 0; i < alive.length; i++) {
-                int live = alive[i];
+                long live = alive[i];
                 if (live == INVALID_LOADER_ID) {
                     return false;
                 }
@@ -342,7 +342,7 @@ final class LoadingConstraints implements ContextAccess {
             return false;
         }
 
-        static Constraint create(int klass, int loader1, int loader2) {
+        static Constraint create(long klass, long loader1, long loader2) {
             if (loader1 == loader2) {
                 return create(klass, loader1);
             }
@@ -352,7 +352,7 @@ final class LoadingConstraints implements ContextAccess {
             return constraint;
         }
 
-        static Constraint create(int klass, int loader) {
+        static Constraint create(long klass, long loader) {
             Constraint constraint = new Constraint(klass);
             constraint.add(loader);
             return constraint;
@@ -365,7 +365,7 @@ final class LoadingConstraints implements ContextAccess {
         return pairings.get(type);
     }
 
-    private int checkConstraint(int klass, Constraint c1) {
+    private long checkConstraint(long klass, Constraint c1) {
         if (c1 != null) {
             if (exists(c1.klass)) {
                 if (exists(klass)) {
@@ -401,7 +401,7 @@ final class LoadingConstraints implements ContextAccess {
         throw meta.throwExceptionWithMessage(meta.java_lang_LinkageError, message);
     }
 
-    private static int getLoaderID(StaticObject loader, Meta meta) {
+    private static long getLoaderID(StaticObject loader, Meta meta) {
         if (StaticObject.isNull(loader)) {
             return meta.getContext().getBootClassLoaderID();
         }
@@ -412,11 +412,11 @@ final class LoadingConstraints implements ContextAccess {
         return classRegistry.getLoaderID();
     }
 
-    private static int getKlassID(Klass k) {
+    private static long getKlassID(Klass k) {
         return k == null ? NULL_KLASS_ID : k.getId();
     }
 
-    private static boolean exists(int klass) {
+    private static boolean exists(long klass) {
         return klass != NULL_KLASS_ID;
     }
