+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/runtime/EspressoContext.java
@@ -158,7 +158,7 @@ public final class EspressoContext {
     // endregion Helpers
 
     // region ID
-    private final int bootClassLoaderID;
+    private final long bootClassLoaderID;
     // endregion ID
 
     // region InitControl
@@ -236,7 +236,7 @@ public final class EspressoContext {
         return logger;
     }
 
-    public int getBootClassLoaderID() {
+    public long getBootClassLoaderID() {
         return bootClassLoaderID;
     }
